Make requestsPerSec accept a Double, close #3844
Motivation:

requestsPerSec currently accept an Int value. It would make sense to accept a Double, all the more as the underlying storage uses Double.

Modification:

Change requestsPerSec type parameter from Int to Double.

Result:

It's now possible to make assertions on requestsPerSec based on Double value

+++ gatling-core/src/main/scala/io/gatling/core/assertion/AssertionBuilders.scala
@@ -25,7 +25,7 @@ class AssertionWithPath(path: AssertionPath)(implicit configuration: GatlingConf
   def allRequests: AssertionWithPathAndCountMetric = new AssertionWithPathAndCountMetric(path, AllRequests)
   def failedRequests: AssertionWithPathAndCountMetric = new AssertionWithPathAndCountMetric(path, FailedRequests)
   def successfulRequests: AssertionWithPathAndCountMetric = new AssertionWithPathAndCountMetric(path, SuccessfulRequests)
-  def requestsPerSec: AssertionWithPathAndTarget[Int] = new AssertionWithPathAndTarget[Int](path, MeanRequestsPerSecondTarget)
+  def requestsPerSec: AssertionWithPathAndTarget[Double] = new AssertionWithPathAndTarget[Double](path, MeanRequestsPerSecondTarget)
 }
 
 class AssertionWithPathAndTimeMetric(path: AssertionPath, metric: TimeMetric)(implicit configuration: GatlingConfiguration) {
